Added int layer orders and order_items

+++ models/staging/stg_retail__order_items.sql
@@ -7,6 +7,6 @@ order_item_id,
 order_item_order_id AS order_id,
 order_item_product_id AS product_id,
 order_item_quantity AS quantity,
-order_item_subtotal AS subtotal,
-order_item_product_price AS product_price 
+order_item_subtotal::NUMERIC AS subtotal,
+order_item_product_price::NUMERIC AS product_price 
 FROM order_items
